requirements.txt
asgiref==3.3.1
attrs==20.3.0
braintree==4.5.0
certifi==2020.12.5
chardet==4.0.0
Django==3.0.8
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
